refactor: 임시핀 업데이트 query 토드 분리
