+++ src/pages/pinCreate/PinCreate.tsx
@@ -18,13 +18,14 @@ import { useTitle } from '@/hooks/pin/useTitle';
 import { useImageUpload } from '@/hooks/pin/useImageUpload';
 import { useLink } from '@/hooks/pin/useLink';
 import { useExplain } from '@/hooks/pin/useExplain';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import { getTempsPinCheck } from '@/services/getTempsPinCheck';
-import {
-  putTempPinEdit,
-  PutTempPinEditParams,
-} from '@/services/putTempPinEdit';
+// import {
+//   putTempPinEdit,
+//   PutTempPinEditParams,
+// } from '@/services/putTempPinEdit';
 import TempPinEditModal from '@/components/@Modal/tempPinEdit/TempPinEditModal';
+import useTempPinUpdate from '@/hooks/queries/useTempPinUpdate';
 
 const PinCreate = () => {
   const {
@@ -41,29 +42,7 @@ const PinCreate = () => {
     queryFn: () => getTempsPinCheck('taewoktest'),
   });
 
-  // 현재 선택중인 임시핀에 대한 정보를 업데이트
-  useEffect(() => {
-    if (pinList) {
-      const matchingPins = pinList.filter((pin) =>
-        selectPinList.some((selectPin) => selectPin.tempPinNo === pin.tempPinNo)
-      );
-      setSelectPinList(matchingPins);
-    }
-  }, [pinList]);
-
-  const queryClient = useQueryClient();
-
-  const { mutate: tempPinUpdate } = useMutation({
-    mutationFn: ({ editData, pinNo }: PutTempPinEditParams) =>
-      putTempPinEdit({ editData, pinNo }),
-    onSuccess: () => {
-      // 성공시 임시핀 목록 useQuery 업데이트
-      queryClient.invalidateQueries({ queryKey: ['tempPinList'] });
-    },
-    onError: (err) => {
-      console.log(err);
-    },
-  });
+  const { mutate: tempPinUpdate } = useTempPinUpdate();
 
   /** 모든 임시핀을 선택 함수 */
   const allPinSelect = () => {
@@ -77,15 +56,6 @@ const PinCreate = () => {
   });
   const { link, linkOnChange, linkReset } = useLink();
   const { explain, explainOnChange, explainReset, textareaRef } = useExplain();
-
-  /** 핀의 input 내용을 모두 reset하는 함수 */
-  const pinFormReset = () => {
-    titleReset();
-    imageReset();
-    linkReset();
-    explainReset();
-    optionReset();
-  };
 
   const { tagSearch, tagList, tagSearchOnChange, selectTagDelet, tagReset } =
     useTagSearch();
@@ -109,6 +79,17 @@ const PinCreate = () => {
     optionReset,
   } = useOptionSettings();
 
+  const { isModalOpen } = useModalStore();
+
+  /** 핀의 input 내용을 모두 reset하는 함수 */
+  const pinFormReset = () => {
+    titleReset();
+    imageReset();
+    linkReset();
+    explainReset();
+    optionReset();
+  };
+
   // 임시핀 내용 수정 useEffect
   useEffect(() => {
     const editData = {
@@ -121,10 +102,17 @@ const PinCreate = () => {
     if (currentPin) tempPinUpdate({ editData, pinNo: currentPin.tempPinNo });
   }, [explain, title, link, isComment, currentPin, tempPinUpdate]);
 
-  const { isModalOpen } = useModalStore();
+  // 현재 선택중인 임시핀에 대한 정보를 업데이트
+  useEffect(() => {
+    if (pinList) {
+      const matchingPins = pinList.filter((pin) =>
+        selectPinList.some((selectPin) => selectPin.tempPinNo === pin.tempPinNo)
+      );
+      setSelectPinList(matchingPins);
+    }
+  }, [pinList]);
 
   const scrollRef = useRef<HTMLDivElement | null>(null); // scrollRef의 타입은 HTMLElement | null
-
   useEffect(() => {
     // 스크롤 숨기기
     document.documentElement.style.overflow = 'hidden';
@@ -144,8 +132,17 @@ const PinCreate = () => {
           selectPinList.length > 0 && 'opacity-25 pointer-events-none'
         } flex-grow pb-8 overflow-scroll max-h-[100vh]`}
       >
-        <h1 className="flex items-center h-[74.31px] border-b-[1px] pl-4 text-xl font-semibold">
-          핀 만들기
+        <h1 className="justify-between flex items-center h-[74.31px] border-b-[1px] pl-4 text-xl font-semibold">
+          <span>핀 만들기</span>
+          {imgPreview && (
+            <div className="pr-3">
+              <Button
+                color="red"
+                text="게시"
+                className="w-[64px] h-[48px] text-[16px] "
+              />
+            </div>
+          )}
         </h1>
         <div className="mx-2 min-w-[584px] flex flex-col lg:gap-12 lg:flex-row lg:justify-center">
           <div className="py-4 mt-4 flex flex-col items-center">
